# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Course Planner API
  version: 1.0.0
  description: |
    **Backend REST API untuk Sistem Perencanaan Perkuliahan**
    
    API ini menyediakan layanan lengkap untuk pengelolaan data akademik termasuk manajemen mata kuliah, kelas, ruangan, dosen, dan Kartu Rencana Studi (KRS) mahasiswa.
    
    ---
    
    ## ⚡ Tech Stack
    
    - **Framework:** Go Fiber v2
    - **Database:** PostgreSQL + GORM
    - **Authentication:** JWT (JSON Web Token)
    - **Documentation:** OpenAPI 3.0
    
    ---
    
    ## 🔐 Authentication
    
    API ini menggunakan **Bearer Token (JWT)** untuk autentikasi. Setelah login, sertakan token di header:
    
    ```
    Authorization: Bearer <your_token>
    ```
    
    ---
    
    ## 📋 Available Modules
    
    | Module | Description |
    |--------|-------------|
    | 🔑 **Auth** | Register & Login user |
    | 📚 **Courses** | CRUD mata kuliah |
    | 🏫 **Classes** | Manajemen kelas perkuliahan |
    | 🚪 **Rooms** | Manajemen ruangan |
    | 👨‍🏫 **Dosen** | Data dosen pengajar |
    | 📝 **KRS** | Kartu Rencana Studi mahasiswa |
    | 📚 **Books** | Referensi buku (Google Books API) |
    
    ---
    
    ## 👥 Development Team
    
    | Developer | GitHub |
    |-----------|--------|
    | **Daffa Agustian Saadi** | [@cikinodapz](https://github.com/cikinodapz) |
    | **Nadia Deari Hanifah** | [@nadia1624](https://github.com/nadia1624) |
...
    
    ---
    
...
  license:
    name: MIT License
...
servers:
...
paths:
  /api/auth/register:
    post:
      summary: Register user baru
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registrasi berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: registration successful
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/login:
    post:
      summary: Login dan dapatkan JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/me/:
    get:
      summary: Contoh endpoint terproteksi
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: authenticated endpoint
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/classes:
    get:
      summary: Daftar kelas (admin)
      tags: [Classes]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar kelas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Buat kelas baru (admin)
      tags: [Classes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassRequest'
      responses:
        '201':
          description: Kelas dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/krs:
    get:
      summary: Lihat KRS aktif (mahasiswa)
      tags: [KRS]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: KRS ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/KRS'
                  semester:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: KRS belum ada
  /api/krs/available-classes:
    get:
      summary: Daftar kelas yang tersedia (mahasiswa)
      tags: [KRS]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar kelas tersedia
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Class'
                  semester:
                    type: string
  /api/krs/items:
    post:
      summary: Ambil kelas (mahasiswa)
      tags: [KRS]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                class_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Kelas diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Matakuliah berhasil ditambahkan ke KRS
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Konflik (kelas sudah ada / sudah diverifikasi)
...
  /api/dosen/students:
    get:
      summary: Daftar mahasiswa bimbingan (dosen PA)
      description: Hanya untuk role `dosen`.
      tags: [Dosen PA]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar mahasiswa bimbingan
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/dosen/students/{mahasiswaId}/krs:
    get:
      summary: Lihat KRS mahasiswa bimbingan (semester berjalan)
      description: Hanya untuk role `dosen`.
      tags: [Dosen PA]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: mahasiswaId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: KRS mahasiswa
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/KRS'
                  semester:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: KRS tidak ditemukan
...
  /api/admin/courses:
    get:
      summary: Daftar semua mata kuliah (admin)
      tags: [Admin - Courses]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar mata kuliah
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Buat mata kuliah baru (admin)
      tags: [Admin - Courses]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Mata kuliah dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Course created successfully
                  course:
                    $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/rooms:
    get:
      summary: Daftar semua ruangan (admin)
      tags: [Admin - Rooms]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar ruangan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Buat ruangan baru (admin)
      tags: [Admin - Rooms]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Ruangan dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Room created successfully
                  room:
                    $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/dosen:
    get:
      summary: Daftar semua dosen (admin)
      tags: [Admin - Dosen]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar dosen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dosen'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/books:
    get:
      summary: Cari buku dari Google Books API
      description: |
        **[UAS Feature]** Endpoint ini mengkonsumsi Google Books API (External Public API).
        
        - **Autentikasi External API**: API Key (GOOGLE_BOOKS_API_KEY)
        - **Provider**: Google Books API (https://www.googleapis.com/books/v1/volumes)
        
        Gunakan endpoint ini untuk mencari buku referensi berdasarkan kata kunci.
      tags: [Books (External API)]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: query
          required: true
          description: Kata kunci pencarian (contoh algoritma, pemrograman)
          schema:
            type: string
            example: algoritma
        - in: query
          name: maxResults
          required: false
          description: Jumlah maksimal hasil pencarian (default 10, max 40)
          schema:
            type: integer
            default: 10
            maximum: 40
      responses:
        '200':
          description: Daftar buku ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Error dari Google Books API
  /api/books/{id}:
    get:
      summary: Detail buku dari Google Books API
      description: |
        **[UAS Feature]** Endpoint ini mengkonsumsi Google Books API (External Public API).
        
        - **Autentikasi External API**: API Key (GOOGLE_BOOKS_API_KEY)
        - **Provider**: Google Books API (https://www.googleapis.com/books/v1/volumes)
        
        Gunakan endpoint ini untuk mendapatkan detail lengkap buku berdasarkan Volume ID dari Google Books.
      tags: [Books (External API)]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Google Books Volume ID
          schema:
            type: string
            example: zyTCAlFPjgYC
      responses:
        '200':
          description: Detail buku
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Buku tidak ditemukan
        '500':
          description: Error dari Google Books API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Monkey D. Luffy
        email:
          type: string
          format: email
          example: luffy@example.com
        password:
          type: string
          format: password
          example: strawhat123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 4f2c0f4e-d0b4-4a67-9f2e-24ce01234567
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          example: mahasiswa
        nim:
          type: string
        nidn:
          type: string
        dosen_pa_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        kode:
          type: string
          example: IF101
        nama:
          type: string
          example: Algoritma dan Pemrograman
        sks:
          type: integer
          example: 3
    CreateCourseRequest:
      type: object
      required: [kode, nama, sks]
      properties:
        kode:
          type: string
          example: IF101
        nama:
          type: string
          example: Algoritma dan Pemrograman
        sks:
          type: integer
          example: 3
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        nama:
          type: string
          example: Lab Komputer A
    CreateRoomRequest:
      type: object
      required: [nama]
      properties:
        nama:
          type: string
          example: Lab Komputer A
...
    Dosen:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440002
        name:
          type: string
          example: Dr. John Doe
        email:
          type: string
          format: email
          example: johndoe@university.ac.id
        role:
          type: string
          example: dosen
        nidn:
          type: string
          example: "0012345678"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Class:
      type: object
      properties:
        id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        dosen_id:
          type: string
          format: uuid
        nama_kelas:
          type: string
        hari:
          type: string
          example: Senin
        jam_mulai:
          type: string
          example: "08:00"
        jam_selesai:
          type: string
          example: "10:00"
        room_id:
          type: string
          format: uuid
        kuota:
          type: integer
        semester_penawaran:
          type: string
          example: ganjil
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateClassRequest:
      type: object
      required:
        [course_id, dosen_id, nama_kelas, hari, jam_mulai, jam_selesai, room_id, kuota, semester_penawaran]
      properties:
        course_id:
          type: string
          format: uuid
        dosen_id:
          type: string
          format: uuid
        nama_kelas:
          type: string
        hari:
          type: string
        jam_mulai:
          type: string
          example: "08:00"
        jam_selesai:
          type: string
          example: "10:00"
        room_id:
          type: string
          format: uuid
        kuota:
          type: integer
        semester_penawaran:
          type: string
          example: ganjil
...
    KRS:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mahasiswa_id:
          type: string
          format: uuid
        semester:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/KRSItem'
    KRSItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        status:
          type: string
        class:
          $ref: '#/components/schemas/Class'

    # Books Schemas (UAS - External API)
    BookSearchResponse:
      type: object
      description: Response dari pencarian buku Google Books API
      properties:
        total_items:
          type: integer
          description: Total buku yang ditemukan
          example: 150
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookBrief'
    BookBrief:
      type: object
      description: Ringkasan informasi buku
      properties:
        id:
          type: string
          description: Google Books Volume ID
          example: zyTCAlFPjgYC
        title:
          type: string
          example: Introduction to Algorithms
        authors:
          type: array
          items:
            type: string
          example: ["Thomas H. Cormen", "Charles E. Leiserson"]
        publisher:
          type: string
          example: MIT Press
        published_date:
          type: string
          example: "2009"
        description:
          type: string
          example: "Buku panduan lengkap tentang algoritma..."
        thumbnail:
          type: string
          format: uri
          example: https://books.google.com/books/content?id=xxx
        info_link:
          type: string
          format: uri
          example: https://books.google.com/books?id=xxx
        isbn_10:
          type: string
          example: "0262033844"
        isbn_13:
          type: string
          example: "9780262033848"
    BookDetail:
      type: object
      description: Detail lengkap informasi buku
      properties:
        id:
          type: string
          description: Google Books Volume ID
          example: zyTCAlFPjgYC
        title:
          type: string
          example: Introduction to Algorithms
        subtitle:
          type: string
          example: Third Edition
        authors:
          type: array
          items:
            type: string
          example: ["Thomas H. Cormen", "Charles E. Leiserson"]
        publisher:
          type: string
          example: MIT Press
        published_date:
          type: string
          example: "2009-07-31"
        description:
          type: string
          example: "This book covers a broad range of algorithms in depth..."
        page_count:
          type: integer
          example: 1312
        categories:
          type: array
          items:
            type: string
          example: ["Computers", "Programming"]
        average_rating:
          type: number
          format: float
          example: 4.5
        ratings_count:
          type: integer
          example: 127
        language:
          type: string
          example: en
        thumbnail:
          type: string
          format: uri
        preview_link:
          type: string
          format: uri
        info_link:
          type: string
          format: uri
        isbn_10:
          type: string
          example: "0262033844"
        isbn_13:
          type: string
          example: "9780262033848"